requirements.txt
# Implicit dependencies (optional dependencies of dependencies)
social-auth-app-django==4.0.0
python-social-auth==0.3.6
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0

# Explicit dependencies (references in code)
Django==2.2.16
django-crispy-forms==1.9.2
django-model-utils==4.0.0
django-munin==0.2.1
python-memcached==1.59
lxml==4.5.2
factory-boy==2.12.0
geoip2==4.1.0
Pillow==7.2.0
GitPython==3.1.9
easy-thumbnails==2.7
beautifulsoup4==4.9.3
django-recaptcha==2.0.6
google-api-python-client==1.12.3
toml==0.10.1
requests==2.24.0
homoglyphs==2.0.3

# Api dependencies
djangorestframework==3.12.1
djangorestframework-xml==2.0.0
django-filter==2.4.0
django-oauth-toolkit==1.3.2
drf-extensions==0.6.0
django-rest-swagger==2.2.0
django-cors-middleware==1.5.0
dry-rest-permissions==0.1.10
oauth2client==4.1.3

# Zep 12 dependency
# next versions of this libraries break previous behavior for slug generation from string with single quotes
django-uuslug==1.0.3
python-slugify==1.1.4
